+++ src/pages/NotFoundPage/NotFoundPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import icon404 from '../../assets/images/icon-404.svg';
 import Button from '../../components/common/Button/ButtonContainer';
-import { Section, Image, ErrorMessage } from './style/NotFoundPageStyle';
+import { Section, Image, ErrorMessage } from './NotFoundPageStyle';
 
 const NotFound = () => {
   const navigate = useNavigate();
